refacting keras GraphSearchUtils
