+++ TrainingExtensions/tensorflow/src/python/aimet_tensorflow/keras/cross_layer_equalization.py
@@ -37,7 +37,6 @@
 # =============================================================================
 """Cross Layer Equalization"""
 
-import collections
 import typing
 
 import numpy as np
@@ -46,7 +45,7 @@ import tensorflow as tf
 import aimet_common.libpymo as libpymo
 from aimet_common.utils import AimetLogger
 from aimet_tensorflow.keras.batch_norm_fold import fold_all_batch_norms
-from aimet_tensorflow.keras.connectedgraph import ConnectedGraph
+from aimet_tensorflow.keras.graphsearchtuils import GraphSearchUtils, ClsSet
 from aimet_tensorflow.keras.utils import model_transform_utils
 from aimet_tensorflow.keras.utils.weight_tensor_utils import WeightTensorUtils
 
@@ -57,19 +56,8 @@ BatchNormFoldedPair = typing.Union[typing.Tuple[tf.keras.layers.Conv2D,
                                    typing.Tuple[tf.keras.layers.Dense,
                                                 tf.keras.layers.BatchNormalization]]
 
-ClsSet = typing.Union[typing.Tuple[tf.keras.layers.Conv2D,
-                                   tf.keras.layers.Conv2D],
-                      typing.Tuple[tf.keras.layers.Conv2D,
-                                   tf.keras.layers.DepthwiseConv2D,
-                                   tf.keras.layers.Conv2D]]
-
 ScaleFactor = typing.Union[np.ndarray, typing.Tuple[np.ndarray, np.ndarray]]
 ReluFlag = typing.Union[bool, typing.Tuple[bool, bool]]
-
-cls_supported_layers = (tf.keras.layers.Conv2D, tf.keras.layers.Conv1D)
-zero_padding_layers = (tf.keras.layers.ZeroPadding2D, tf.keras.layers.ZeroPadding1D)
-cls_supported_activations = (tf.keras.layers.ReLU, tf.keras.layers.PReLU)
-
 
 class ClsSetInfo:
     """
@@ -106,7 +94,6 @@ class ClsSetInfo:
     def __init__(self, cls_pair_1: ClsSetLayerPairInfo, cls_pair_2: ClsSetLayerPairInfo = None):
         """
         Constructor takes 2 pairs if Depth-wise separable layer is being folded
-
         :param cls_pair_1: Pair between two conv or conv and depth-wise conv
         :param cls_pair_2: Pair between depth-wise conv and point-wise conv
         """
@@ -120,265 +107,6 @@ class ClsSetInfo:
             return self.cls_pair_info_list == other.cls_pair_info_list
 
         return False
-
-
-class GraphSearchUtils:
-    """Implements graph search utils required by CLE feature"""
-
-    def __init__(self, model: tf.keras.Model):
-        """
-        :param model: Keras Model that is built (Sequential, Functional)
-        """
-        self._connected_graph = ConnectedGraph(model)
-        self._ordered_module_list = self._get_ordered_list_of_conv_modules()
-
-    def _get_ordered_list_of_conv_modules(self):
-        """
-        Finds order of nodes in graph
-
-        :return: List of name, layer tuples in graph in order
-        """
-        result = []
-        for op in self._connected_graph.ordered_ops:
-            layer = op.get_module()
-            if isinstance(layer, cls_supported_layers):
-                result.append([layer.name, layer])
-
-        return result
-
-    def find_layer_groups_to_scale(self) -> typing.List[typing.List[tf.keras.layers.Conv2D]]:
-        """
-        Find layer groups to scale
-
-        :return: List of groups of layers. Each group can be independently equalized
-        """
-        # Find the input node(s) in the graph
-        input_nodes = []
-        for op in self._connected_graph.get_all_ops().values():
-            if op.inputs and op.inputs[0].is_model_input:
-                input_nodes.append(op)
-
-        layer_groups = []
-        for op in input_nodes:
-            self.find_downstream_layer_groups_to_scale(op, layer_groups)
-
-        # Sort the layer groups in order of occurrence in the model
-        ordered_layer_groups = []
-        for _, module in self._ordered_module_list:
-            for layer_group in layer_groups:
-                if layer_group[0] is module:
-                    ordered_layer_groups.append(layer_group)
-
-        return ordered_layer_groups
-
-    @staticmethod
-    def find_downstream_layer_groups_to_scale(op, layer_groups, current_group=None, visited_nodes=None):
-        """
-        Recursive function to find cls layer groups downstream from a given op
-
-        :param op: Starting op to search from
-        :param layer_groups: Running list of layer groups
-        :param current_group: Running current layer group
-        :param visited_nodes: Running list of visited nodes (to short-circuit recursion)
-        :return: None
-        """
-
-        if not visited_nodes:
-            visited_nodes = []
-        if not current_group:
-            current_group = []
-
-        if op in visited_nodes:
-            return
-        visited_nodes.append(op)
-
-        current_layer = op.get_module()
-        # Conv2D, Conv1D or its subclass is added to the current group
-        if current_layer and isinstance(current_layer, cls_supported_layers):
-            current_group.append(current_layer)
-
-        # Terminating condition for current group
-        if not current_layer or not GraphSearchUtils._is_supported_layer_case(current_layer):
-            if (len(current_group) > 1) and (current_group not in layer_groups):
-                layer_groups.append(current_group)
-            current_group = []
-
-        if op.output:
-            for consumer in op.output.consumers:
-                GraphSearchUtils.find_downstream_layer_groups_to_scale(consumer, layer_groups,
-                                                                       current_group, visited_nodes)
-
-        # Reached a leaf.. See if the current group has something to grab
-        if (len(current_group) > 1) and (current_group not in layer_groups):
-            layer_groups.append(current_group)
-
-    @staticmethod
-    def _is_supported_layer_case(layer: tf.keras.layers.Layer) -> bool:
-        """
-        Check if the current layer is CLS supported layers or a supported activation layer
-
-        :param layer: tf.keras.layers.Layer
-        :return: True if it's CLS supported layers or a supported layer
-        """
-        return isinstance(layer, (cls_supported_layers + zero_padding_layers)) or \
-            GraphSearchUtils._is_supported_activations(layer) or \
-            GraphSearchUtils.is_folded_batch_normalization(layer)
-
-    @staticmethod
-    def is_folded_batch_normalization(layer: tf.keras.layers.Layer) -> bool:
-        """
-        Method to check if layer is folded batchnorm or not
-
-        :param layer: layer to check if it is folded batch norm
-        :return: True if it is folded batch norm, False if not
-        """
-        if not isinstance(layer, tf.keras.layers.BatchNormalization):
-            return False
-
-        return np.all(layer.beta == 0.0) and np.all(layer.gamma == 1.0)
-
-    @staticmethod
-    def _is_supported_activations(layer: tf.keras.layers.Layer) -> bool:
-        """
-        Check if the current layer is a supported activation layer
-
-        :param layer: tf.keras.layers.Layer
-        :return: True if layer is ReLU, PReLU or Activation with supported type
-        """
-        # Case of explicit layer such as tf.keras.layers.ReLU
-        if isinstance(layer, cls_supported_activations):
-            return True
-
-        # Case of implicit layer such as tf.keras.layers.Activation(tf.nn.relu)
-        # Note: PReLU is not supported by implicit approach until TF 2.4
-        layer_config = layer.get_config()
-        activation = layer_config.get("activation")
-
-        if activation is None:
-            return False
-
-        return activation in ["relu", "relu6"]
-
-    @staticmethod
-    def convert_layer_group_to_cls_sets(layer_group: typing.List[tf.keras.layers.Conv2D]) \
-            -> typing.List[ClsSet]:
-        """
-        Helper function to convert a layer group to a list of cls sets
-        :param layer_group: Given layer group to convert
-        :return: List of cls sets
-        """
-        cls_sets = []
-
-        layer_group = collections.deque(layer_group)
-        prev_layer_to_scale = layer_group.popleft()
-        while layer_group:
-            next_layer_to_scale = layer_group.popleft()
-
-            if isinstance(next_layer_to_scale, tf.keras.layers.DepthwiseConv2D):
-                next_non_depthwise_conv_layer = layer_group.popleft()
-                # DepthwiseConv layer right after DepthwiseConv layer is not currently supported
-                if isinstance(next_non_depthwise_conv_layer, tf.keras.layers.DepthwiseConv2D):
-                    _logger.error("Consecutive DepthwiseConv layer not currently supported")
-                    raise NotImplementedError("Consecutive DepthwiseConv layer not currently supported")
-
-                cls_sets.append(
-                    (prev_layer_to_scale, next_layer_to_scale, next_non_depthwise_conv_layer))
-                prev_layer_to_scale = next_non_depthwise_conv_layer
-            else:
-                cls_sets.append((prev_layer_to_scale, next_layer_to_scale))
-                prev_layer_to_scale = next_layer_to_scale
-
-        return cls_sets
-
-    @staticmethod
-    def is_relu_activation_present_in_cls_sets(cls_sets: typing.List[ClsSet]) \
-            -> typing.List[typing.Union[bool, typing.Tuple[bool, bool]]]:
-        """
-        Check if there is ReLU or PReLU activation between cls sets
-        :param cls_sets: List of ClsSet to find ReLU activation in
-        :return: List of ReLU activation preset flags (bool or tuple of bool) corresponding to input cls_sets param
-        """
-
-        is_relu_activation_in_cls_sets = []
-        for cls_set in cls_sets:
-            cls_set = cls_set[:-1]
-
-            is_relu_activation_in_cls_set = []
-            for layer in cls_set:
-                has_relu_activation = GraphSearchUtils._does_layer_have_relu_activation(layer)
-                is_relu_activation_in_cls_set.append(has_relu_activation)
-
-            if len(is_relu_activation_in_cls_set) == 1:
-                is_relu_activation_in_cls_sets.append(is_relu_activation_in_cls_set[0])
-            else:
-                is_relu_activation_in_cls_sets.append(tuple(is_relu_activation_in_cls_set))
-
-        return is_relu_activation_in_cls_sets
-
-    @staticmethod
-    def _does_layer_have_relu_activation(layer: tf.keras.layers.Conv2D) -> bool:
-        """
-        Check if layer has ReLU or PReLU activation function
-        :param layer: Conv2D or it's subclass to check activation function
-        :return: True If layer has ReLU or PReLU activation, otherwise False
-        """
-        def _get_activation_type(_layer: tf.keras.layers.Layer) -> str:
-            """
-            Get activation name string from _layer
-            :param _layer: tf.keras.layers.Layer
-            :return: activation name string
-            """
-            activation_info = tf.keras.activations.serialize(_layer.activation)
-
-            if isinstance(activation_info, str):
-                # Instantiating like tf.keras.layers.Conv2D(8, kernel_size=3, activation=tf.keras.activations.relu)
-                #   has the result of serialization as str type
-                _activation_type = activation_info
-            elif isinstance(activation_info, dict):
-                # Instantiating like tf.keras.layers.Conv2D(8, kernel_size=3, activation=tf.keras.layers.ReLU())
-                #   has the result of serialization as dict type
-                _activation_type = activation_info["class_name"].lower()
-            else:
-                raise NotImplementedError("Not supported format")
-
-            return _activation_type
-
-        def _get_outbound_layer(_layer: tf.keras.layers.Layer) -> tf.keras.layers.Layer:
-            """
-            Get outbound layer from current layer
-            If right after layer is folded batchnorm, find next applicable layer
-            :param _layer: tf.keras.layers.Layer
-            :return: outbound layer of _layer
-            """
-            assert len(_layer.outbound_nodes) == 1
-
-            _outbound_layer = _layer.outbound_nodes[0].outbound_layer
-            return _get_outbound_layer(_outbound_layer) \
-                if GraphSearchUtils.is_folded_batch_normalization(_outbound_layer) \
-                else _outbound_layer
-
-        activation_type = _get_activation_type(layer)
-        supported_activation_types = {"relu", "prelu"}
-
-        # If activation parameter is not set or None, default activation_type is linear
-        if activation_type == "linear" and layer.outbound_nodes:
-            outbound_layer = _get_outbound_layer(layer)
-
-            # Case 1. Non-fused use case
-            # Case 1-1. Conv(..., activation=None) -> ReLU() or
-            #           Conv(..., activation=None) -> Folded BN -> ReLU()
-            is_using_relu_layer = isinstance(outbound_layer, cls_supported_activations)
-
-            # Case 1-2. Conv(..., activation=None) -> Activation(activation="relu") or
-            #           Conv(..., activation=None) -> Folded BN -> Activation(activation="relu")
-            is_using_activation_layer = isinstance(outbound_layer, tf.keras.layers.Activation) and \
-                                        _get_activation_type(outbound_layer) in supported_activation_types
-            return is_using_relu_layer or is_using_activation_layer
-
-        # Case 2. Fused use case
-        # e.g., Conv(..., activation="relu") or Conv(..., activation=tf.keras.layers.PReLU())
-        return activation_type in supported_activation_types
-
 
 class CrossLayerScaling:
     """
@@ -532,7 +260,6 @@ class CrossLayerScaling:
             typing.List[typing.Union[np.ndarray, typing.Tuple[np.ndarray, np.ndarray]]]:
         """
         Scale each cls set
-
         :param cls_sets: Cls sets to scale
         :return: List of scale factors corresponding to each scaled cls set
         """
@@ -551,7 +278,6 @@ class CrossLayerScaling:
                                  is_relu_activation_in_cls_sets: typing.List[ReluFlag]) -> typing.List[ClsSetInfo]:
         """
         Binds information from there separate lists into one [ClsInfoSet] data structure
-
         :param cls_sets: List of CLS sets
         :param scale_factors: List of scale factors for each cls set
         :param is_relu_activation_in_cls_sets: List of ReLU flag whether there is ReLU activation in each cls set
@@ -592,7 +318,6 @@ class CrossLayerScaling:
     def scale_model(model: tf.keras.Model) -> typing.List[ClsSetInfo]:
         """
         Uses cross-layer scaling to scale all applicable layers in the given model
-
         :param model: tf.keras.Model
         :return: CLS information for each CLS set
         """
@@ -631,7 +356,6 @@ class HighBiasFold:
                   bn_layers: typing.Dict[tf.keras.layers.Conv2D, tf.keras.layers.BatchNormalization]):
         """
         Folds bias values greater than 3 * sigma to next layer's bias
-
         :param cls_set_info_list: List of info elements for each cls set
         :param bn_layers: Key: Conv/Linear layer Value: Corresponding folded BN layer
         """
@@ -663,7 +387,6 @@ class HighBiasFold:
             -> typing.Tuple[libpymo.LayerParams, libpymo.LayerParams]:
         """
         Invokes high bias fold MO API
-
         :param cls_pair_info: Pair of layers that were scaled using CLS and related information
         :param bn_layers: Key: Conv/Linear layer Value: Corresponding folded BN layer
         :return: Updated layer params
@@ -682,7 +405,6 @@ class HighBiasFold:
                                        scaling_parameter: np.ndarray) -> libpymo.BNParamsHighBiasFold:
         """
         Helper method to pack BatchNormalization parameter for high bias fold
-
         :param bn_layer: Target batch normalization layer
         :param scaling_parameter: Scaling parameters for each channel obtained from cross layer scaling
         :return: Packed BNParamsHighBiasFold
@@ -709,7 +431,6 @@ class HighBiasFold:
             -> typing.Tuple[libpymo.LayerParams, libpymo.LayerParams]:
         """
         Helper method to pack information of previous and current layer
-
         :param cls_pair_info: Pair of layers that were scaled using CLS and related information
         :return: Packed layer parameter tuple
         """
@@ -740,7 +461,6 @@ class HighBiasFold:
 def equalize_model(model: tf.keras.Model) -> tf.keras.Model:
     """
     High-level API to perform Cross-Layer Equalization (CLE) on the given model
-
     :param model: tf.keras.Model
     :return: CLE applied tf.keras.Model
     """
@@ -757,7 +477,6 @@ def equalize_bn_folded_model(model: tf.keras.Model,
                              folded_pairs: typing.List[BatchNormFoldedPair]):
     """
     Perform Cross-Layer Scaling (CLS) and High Bias Folding (HBF) on a batchnorm-folded model in-place
-
     :param model: BatchNorm-folded model to equalize
     :param folded_pairs: List of pairs of folded layers
     """
